+++ ccd/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "0.29.0"
+__version__ = "0.30.0"
 
 __all__ = ["__version__"]